requirements.txt
***REMOVED***
litellm>=1.40.0
openai>=1.30.0

# Server
fastapi>=0.115.0
uvicorn[standard]>=0.29.0

# CLI
typer>=0.12.0
rich>=13.7.0

# Models + Config
pydantic>=2.7.0
pydantic-settings>=2.3.0
pyyaml>=6.0.1

# Templates
jinja2>=3.1.4

# Caching
cachetools>=5.3.3

# Logging
colorlog>=6.8.2

# HTTP clients
httpx>=0.27.0
requests>=2.32.0

# Slack
slack-bolt>=1.18.1

# PDF
weasyprint>=61.0
markdown>=3.6

# CLI interactive
prompt-toolkit>=3.0.47

# Toolset deps
kubernetes>=30.1.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
